check for dulicate data & not found issues

+++ app/Http/Controllers/ServiceController.php
@@ -21,20 +21,41 @@ class ServiceController extends Controller
 
     public function store(ServiceRequest $request): JsonResponse
     {
-        $service = $this->service->store($request->validated());;
+        $existingService = $this->service->findByName($request->name);
+
+        if ($existingService) {
+            return response()->json(['message' => 'Service already exists.'], 409);
+        }
+
+        $service = $this->service->store($request->validated());
         return response()->json($service, 201);
     }
 
     public function update(ServiceRequest $request, $id): JsonResponse
     {
         $service = $this->service->find($id);
+
+        if (!$service) {
+            return response()->json(['message' => 'Service not found.'], 404);
+        }
+
+        $existingService = $this->service->findByName($request->input('name'));
+
+        if ($existingService) {
+            return response()->json(['message' => 'Service with this name already exists.'], 409);
+        }
+
         $updated = $service->update($request->validated());
+
         return response()->json($updated);
     }
 
     public function destroy($id): JsonResponse
     {
         $service = $this->service->find($id);
+        if (!$service) {
+            return response()->json(['message' => 'Service not found.'], 404);
+        }
         $this->service->destroy($service);
 
         return response()->json(['message' => 'Service deleted successfully.']);
